edited form category,model category, views category,templates category. added alerts partials and edit settings for django compressor

+++ category/views.py
@@ -6,12 +6,17 @@ from .forms import CategoryForm
 
 def delete_category(request,id):
     category = get_object_or_404(Category,pk=id)
+
+    if category.user != request.user:
+        messages.error(request, "Você não tem permissão para deletar esta categoria.")
+        return redirect('categories')
+    
     category.delete()
     messages.success(request,f'Categoria {category.name} deletada com sucesso')
     return redirect('categories')
 
 def get_page_categories(request):
-    categories = Category.objects.all().order_by('-id')
+    categories = Category.objects.filter(user=request.user).order_by('-id')
     template_data = {'categories': categories}
 
     if request.method == 'POST':
@@ -19,9 +24,14 @@ def get_page_categories(request):
 
         if id_category:  
             category = get_object_or_404(Category, pk=id_category)
-            form = CategoryForm(request.POST, instance=category)
+
+            if category.user != request.user:
+                messages.error(request, "Você não tem permissão para Atualizar esta categoria.")
+                return redirect('categories')
+            
+            form = CategoryForm(request.POST, instance=category,user=request.user)
         else: 
-            form = CategoryForm(request.POST)
+            form = CategoryForm(request.POST,user=request.user)
 
         if form.is_valid():
             category_instance = form.save()
